requirements.txt
Django==1.5.1
MySQL-python==1.2.4
Pillow==2.1.0
South==0.8.1
amqp==1.0.11
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.28
boto==2.9.5
celery-haystack
celery==3.0.19
django-appconf==0.6
django-autofixture==0.3.0
django-celery-transactions==0.1.3
django-celery==3.0.17
django-debug-toolbar==0.9.4
django-haystack==2.0.0
django-picklefield==0.3.0
django-queued-storage==0.6
django-reversion==1.7.1
django-socketio==0.3.0
django-storages==1.1.8
django-tastypie==0.9.15
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
gevent==0.13.8
greenlet==0.4.1
kombu>=2.5.10,<3.0
lxml>3.0
mimeparse==0.1.3
pysolr==3.0.6
python-dateutil==2.1
python-magic==0.4.6
pytz==2013b
requests==1.2.3
six==1.3.0
sphinx-me==0.1.2
wsgiref==0.1.2
django-kombu
kombu-sqlalchemy
django-celery
pysolr
mimeparse
